Reduce amount of copying/cloning within the transformation from `StreamerMessage` to `codec::Block`
We were passing the top element as a reference as well as a bunch of other inner structures. This was causing unecessary cloning of elements but we didn't had ownership of the top-level structure. However this was pretty bad because we actually "own" the `StreamerMessage` which is dropped right after the transformation.

So, everything is now "moved" which means much less data need to be copied/created and mostly everything can be simply be "moved" in the right location. There was also useless usage of `clone` here and there which was doubly not required.